requirements.txt
Django==4.1.1
djangorestframework
django-filter
psycopg2 